% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{oderun}
\alias{oderun}
\title{Run a TB model function}
\usage{
oderun(func, parms = list(), y0 = c(S = 1000, I = 10, R = 10, N = 1020),
  times = 1:1000)
}
\arguments{
\item{func}{a function with signature function(t, y, parms, ...); see
\code{\link[deSolve]{ode}}}

\item{parms}{a list of replacement parameters for the default model
parameters; see \code{\link[deSolve]{ode}}.}

\item{y0}{see \code{\link[deSolve]{ode}}}

\item{times}{see \code{\link[deSolve]{ode}}}
}
\description{
Run a TB model function
}

