% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/TBmodel.R
\name{TBmodel}
\alias{TBmodel}
\title{A function that matches the signature required for ode solvers in deSolve}
\usage{
TBmodel(t, y, parameters)
}
\arguments{
\item{t}{see \code{\link[deSolve]{ode}}}

\item{y}{see \code{\link[deSolve]{ode}}}

\item{parameters}{see \code{\link[deSolve]{ode}}.}
}
\description{
A function that matches the signature required for ode solvers in deSolve
}
\seealso{
\code{\link[deSolve]{ode}}.
}

